implement first part of the new photo editor UI.
Note that it wiped out a lot of features from the page.  We retored the buttons at the bottom.  We will wait to restore the rest at the end of the feature plan, where UI is discussed. We can refer back to the code before this commit for examples of how the UI code should look for implementation.